@@ -10,7 +10,10 @@ repositories {
 
 apply from: '../opencadc.gradle'
 
-sourceCompatibility = 11
+java {
+    sourceCompatibility = JavaVersion.VERSION_21
+    targetCompatibility = JavaVersion.VERSION_21
+}
 
 group = 'org.opencadc'
 
@@ -24,11 +27,19 @@ war {
 }
 
 dependencies {
-    implementation 'org.hibernate.orm:hibernate-core:6.2.0.Final'
+    // Used to query the underlying database.
+    implementation 'org.springframework:spring-context:7.0.5'
+    implementation 'org.springframework:spring-jdbc:7.0.5'
+
     implementation 'org.opencadc:cadc-log:[1.2.1,2.0)'
     implementation 'org.opencadc:cadc-registry:[1.7.3,)'
     implementation 'org.opencadc:cadc-rest:[1.3.14,)'
-    implementation 'org.opencadc:cadc-util:[1.11.6,2.0)'
+
+    implementation 'org.opencadc:cadc-util:[1.12.14,2.0)'
+    constraints {
+        'org.springframework:spring-jdbc:7.0.5'
+    }
+
     implementation 'org.opencadc:cadc-vosi:[1.4.8,2.0)'
     implementation 'org.opencadc:cadc-gms:[1.0.5,2.0)'
 
